requirements.txt
opencv-python
torch
requests
ultralytics
insightface
python-jose
numpy
python-dotenv
scikit-learn